+++ examples/main/shared/viewer/ViewerPlugins.ts
@@ -1,12 +1,21 @@
 import theme from '../theme/theme';
 import { ACTION_BUTTON_TYPE } from 'wix-rich-content-plugin-button/viewer';
+<<<<<<< HEAD:examples/main/shared/viewer/ViewerPlugins.ts
 import { videoTypeMapper, VIDEO_TYPE, pluginVideo } from 'wix-rich-content-plugin-video/viewer';
 import { dividerTypeMapper, pluginDivider } from 'wix-rich-content-plugin-divider/viewer';
 import { htmlTypeMapper, pluginHtml } from 'wix-rich-content-plugin-html/viewer';
 import { soundCloudTypeMapper, pluginSoundCloud } from 'wix-rich-content-plugin-sound-cloud/viewer';
 import { linkTypeMapper, LINK_TYPE, pluginLink } from 'wix-rich-content-plugin-link/viewer';
+=======
+import { VIDEO_TYPE, videoTypeMapper } from 'wix-rich-content-plugin-video/viewer';
+import { dividerTypeMapper } from 'wix-rich-content-plugin-divider/viewer';
+import { htmlTypeMapper } from 'wix-rich-content-plugin-html/viewer';
+import { soundCloudTypeMapper } from 'wix-rich-content-plugin-sound-cloud/viewer';
+import { LINK_TYPE, linkTypeMapper } from 'wix-rich-content-plugin-link/viewer';
+>>>>>>> master:examples/main/shared/viewer/ViewerPlugins.jsx
 import {
   linkPreviewTypeMapper,
+<<<<<<< HEAD:examples/main/shared/viewer/ViewerPlugins.ts
   LINK_PREVIEW_TYPE,
   pluginLinkPreview,
 } from 'wix-rich-content-plugin-link-preview/viewer';
@@ -26,11 +35,24 @@ import {
   verticalEmbedTypeMapper,
   pluginVerticalEmbed,
 } from 'wix-rich-content-plugin-vertical-embed/viewer';
+=======
+} from 'wix-rich-content-plugin-link-preview/viewer';
+import { imageTypeMapper } from 'wix-rich-content-plugin-image/viewer';
+import { tableTypeMapper } from 'wix-rich-content-plugin-table/viewer';
+
+import { galleryTypeMapper, GALLERY_TYPE } from 'wix-rich-content-plugin-gallery/viewer';
+import { mapTypeMapper } from 'wix-rich-content-plugin-map/viewer';
+import { giphyTypeMapper, GIPHY_TYPE } from 'wix-rich-content-plugin-giphy/viewer';
+import { buttonTypeMapper } from 'wix-rich-content-plugin-button/viewer';
+import { HashtagDecorator } from 'wix-rich-content-plugin-hashtag/viewer';
+import { verticalEmbedTypeMapper } from 'wix-rich-content-plugin-vertical-embed/viewer';
+>>>>>>> master:examples/main/shared/viewer/ViewerPlugins.jsx
 import {
   createHeadersMarkdownDecorator,
   HEADERS_MARKDOWN_TYPE,
 } from 'wix-rich-content-plugin-headers-markdown';
 import { CodeBlockDecorator } from 'wix-rich-content-plugin-code-block/viewer';
+<<<<<<< HEAD:examples/main/shared/viewer/ViewerPlugins.ts
 import {
   mentionsTypeMapper,
   MENTION_TYPE,
@@ -41,6 +63,10 @@ import {
   pluginFileUpload,
   FILE_UPLOAD_TYPE,
 } from 'wix-rich-content-plugin-file-upload/viewer';
+=======
+import { MENTION_TYPE, mentionsTypeMapper } from 'wix-rich-content-plugin-mentions/viewer';
+import { fileUploadTypeMapper, FILE_UPLOAD_TYPE } from 'wix-rich-content-plugin-file-upload/viewer';
+>>>>>>> master:examples/main/shared/viewer/ViewerPlugins.jsx
 import {
   textColorInlineStyleMapper,
   TEXT_COLOR_TYPE,
@@ -53,7 +79,11 @@ import {
   SpoilerViewerWrapper,
   SPOILER_TYPE,
 } from 'wix-rich-content-plugin-spoiler/viewer';
+<<<<<<< HEAD:examples/main/shared/viewer/ViewerPlugins.ts
 import { accordionTypeMapper, pluginAccordion } from 'wix-rich-content-plugin-accordion/viewer';
+=======
+import { accordionTypeMapper } from 'wix-rich-content-plugin-accordion/viewer';
+>>>>>>> master:examples/main/shared/viewer/ViewerPlugins.jsx
 
 import {
   viewerCustomForegroundStyleFn,
@@ -61,7 +91,11 @@ import {
   viewerCustomBackgroundStyleFn,
 } from '../../src/text-color-style-fn';
 
+<<<<<<< HEAD:examples/main/shared/viewer/ViewerPlugins.ts
 import { pollTypeMapper, pluginPoll, POLL_TYPE } from 'wix-rich-content-plugin-social-polls/viewer';
+=======
+import { pollTypeMapper, POLL_TYPE } from 'wix-rich-content-plugin-social-polls/viewer';
+>>>>>>> master:examples/main/shared/viewer/ViewerPlugins.jsx
 
 import 'wix-rich-content-editor-common/dist/styles.min.css';
 import 'wix-rich-content-common/dist/styles.min.css';
@@ -89,6 +123,7 @@ import 'wix-rich-content-plugin-table/dist/styles.min.css';
 
 import { getBaseUrl } from '../../src/utils';
 import { RichContentViewerProps } from 'wix-rich-content-viewer';
+<<<<<<< HEAD:examples/main/shared/viewer/ViewerPlugins.ts
 import {
   Decorator,
   HASHTAG_TYPE,
@@ -98,6 +133,9 @@ import {
   ViewerPlugin,
   ViewerPluginCreator,
 } from 'wix-rich-content-common';
+=======
+import { Decorator, PluginTypeMapper, RicosContent, UISettings } from 'wix-rich-content-common';
+>>>>>>> master:examples/main/shared/viewer/ViewerPlugins.jsx
 
 const linkPluginSettings = {
   onClick: (event, url) => console.log('link clicked!', url),
@@ -191,6 +229,7 @@ const config: RichContentViewerProps['config'] = {
   },
 };
 
+<<<<<<< HEAD:examples/main/shared/viewer/ViewerPlugins.ts
 export const viewerPlugins: ViewerPlugin[] = [
   pluginVideo(config[VIDEO_TYPE]),
   pluginActionButton(config[ACTION_BUTTON_TYPE]),
@@ -211,6 +250,8 @@ export const viewerPlugins: ViewerPlugin[] = [
   pluginAccordion(),
 ];
 
+=======
+>>>>>>> master:examples/main/shared/viewer/ViewerPlugins.jsx
 export const getConfig = (additionalConfig = {}): RichContentViewerProps['config'] => {
   let _config = { ...config };
   Object.keys(additionalConfig).forEach(key => {