+++ packages/ricos-common/web/src/RicosEngine.tsx
@@ -75,6 +75,7 @@ export class RicosEngine extends Component<EngineProps> {
       mediaSettings = {},
       linkSettings = {},
       linkPanelSettings = {},
+      maxTextLength,
     } = this.props;
 
     const { strategyProps, previewContent, htmls } = this.runStrategies();
@@ -90,6 +91,7 @@ export class RicosEngine extends Component<EngineProps> {
     const isPreview = () => !!(previewContent && !isPreviewExpanded);
     const ricosPropsToMerge: RichContentProps = {
       isMobile,
+      maxTextLength,
       textToolbarType:
         !isMobile && (textToolbarContainer || useStaticTextToolbar) ? 'static' : 'inline',
       config: {